Changed some methods and variables that previously returned ints or floats to doubles

+++ include/gps_wrapper.h
@@ -14,7 +14,7 @@ class GpsWrapper{
     float headingOffset;
     public:
     GpsWrapper(unsigned int gpsDevicePort, float xOffset, float yOffset, float headingOffset = 0);
-    float getHeading();
+    double getHeading();
     wrapper_position getPositions();
